fixed all the tests up
one of them was working as I wanted, but I realized I just gave it the wrong error code. it should be 500 if there's a malformed UUID supplied. I'm also not thinking too much about obscuring information in HTTP codes for a system such as this; I would imagine it is a closed system